Factor VP8 scaffold test helpers
